+++ cub/test/catch2_test_warp_scan_partial_invalid.cu
@@ -58,7 +58,8 @@ struct merge_aggregate_op_t
         thread_data, thread_data, merge_segments_op{error_flag_ptr}, valid_items, warp_aggregate);
     }
 
-    const int tid = cub::RowMajorTid(blockDim.x, blockDim.y, blockDim.z);
+    const int tid =
+      cub::RowMajorTid(static_cast<int>(blockDim.x), static_cast<int>(blockDim.y), static_cast<int>(blockDim.z));
 
     if (tid % LogicalWarpThreads == m_target_thread_id)
     {
@@ -112,7 +113,8 @@ struct merge_init_value_aggregate_op_t
         thread_data, thread_data, initial_value, merge_segments_op{error_flag_ptr}, valid_items, warp_aggregate);
     }
 
-    const int tid = cub::RowMajorTid(blockDim.x, blockDim.y, blockDim.z);
+    const int tid =
+      cub::RowMajorTid(static_cast<int>(blockDim.x), static_cast<int>(blockDim.y), static_cast<int>(blockDim.z));
 
     if (tid % LogicalWarpThreads == m_target_thread_id)
     {
@@ -168,7 +170,7 @@ C2H_TEST(
     tuple_to_segment_op{});
   for (size_t i = 0; i < params::tile_size; i += params::logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   c2h::device_vector<bool> error_flag(1);
@@ -221,7 +223,7 @@ C2H_TEST("Partial warp scan does not apply op to invalid elements and returns va
     tuple_to_segment_op{});
   for (size_t i = 0; i < params::tile_size; i += params::logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   const int target_thread_id = GENERATE_COPY(take(2, random(0, params::logical_warp_threads - 1)));
@@ -284,7 +286,7 @@ C2H_TEST("Partial warp scan does not apply op to invalid elements and works with
     tuple_to_segment_op{});
   for (size_t i = 0; i < params::tile_size; i += params::logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   const type initial_value = segment{0, 1};
@@ -330,7 +332,7 @@ C2H_TEST("Partial warp scan with initial value does not apply op to invalid elem
     tuple_to_segment_op{});
   for (size_t i = 0; i < params::tile_size; i += params::logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   const int target_thread_id = GENERATE_COPY(take(2, random(0, params::logical_warp_threads - 1)));
@@ -383,7 +385,7 @@ C2H_TEST("Partial warp combination scan does not apply op to invalid elements", 
     tuple_to_segment_op{});
   for (size_t i = 0; i < tile_size; i += logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   c2h::device_vector<bool> error_flag(1);
@@ -444,7 +446,7 @@ C2H_TEST("Partial warp combination custom scan does not apply op to invalid elem
     tuple_to_segment_op{});
   for (size_t i = 0; i < tile_size; i += logical_warp_threads)
   {
-    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + i);
+    thrust::copy(in_it, in_it + bounded_valid_items, d_in.begin() + static_cast<std::ptrdiff_t>(i));
   }
 
   const type initial_value = segment{0, 1};
