+++ cub/test/test_allocator.cu
@@ -445,8 +445,8 @@ int main(int argc, char** argv)
   printf("\t CUB CachingDeviceAllocator allocation CPU speedup: %.2f (avg cudaMalloc %.4f ms vs. avg DeviceAllocate "
          "%.4f ms)\n",
          cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis,
-         cuda_malloc_elapsed_millis / timing_iterations,
-         cub_calloc_elapsed_millis / timing_iterations);
+         cuda_malloc_elapsed_millis / static_cast<float>(timing_iterations),
+         cub_calloc_elapsed_millis / static_cast<float>(timing_iterations));
 
   // GPU performance comparisons.  Allocate and free a 1MB block 2000 times
   GpuTimer gpu_timer;
@@ -489,8 +489,8 @@ int main(int argc, char** argv)
   printf("\t CUB CachingDeviceAllocator allocation GPU speedup: %.2f (avg cudaMalloc %.4f ms vs. avg DeviceAllocate "
          "%.4f ms)\n",
          cuda_malloc_elapsed_millis / cub_calloc_elapsed_millis,
-         cuda_malloc_elapsed_millis / timing_iterations,
-         cub_calloc_elapsed_millis / timing_iterations);
+         cuda_malloc_elapsed_millis / static_cast<float>(timing_iterations),
+         cub_calloc_elapsed_millis / static_cast<float>(timing_iterations));
 
   printf("Success\n");
